created gesture module
